+++ bootstrap3/bootstrap.py
@@ -5,7 +5,7 @@ from django.conf import settings
 # Default settings
 BOOTSTRAP3_DEFAULTS = {
     'jquery_url': '//code.jquery.com/jquery.min.js',
-    'base_url': '//netdna.bootstrapcdn.com/bootstrap/3.0.3/',
+    'base_url': '//netdna.bootstrapcdn.com/bootstrap/3.1.1/',
     'css_url': None,
     'theme_url': None,
     'javascript_url': None,