+++ HISTORY.rst
@@ -6,20 +6,22 @@ History
 Work in progress
 ++++++++++++++++
 
-* update to newest Bootstrap (issue #83)
+* show inline form errors in field title (fix issue #81)
+* fixed bugs in demo application
+* update to newest Bootstrap (fix issue #83)
 
 
 2.6.0 (2014-02-20)
 ++++++++++++++++++
 
-* new setting `set_required` to control setting of html required attribute (issue #76)
+* new setting `set_required` to control setting of HTML `required` attribute (fix issue #76)
 
 
 2.5.6 (2014-01-23)
 ++++++++++++++++++
 
 * project refactored
-* added skeleton for creating documentation to fix issue #30
+* added skeleton for creating documentation (fix issue #30)
 * fixed FileField issues