+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 Work in progress
 ++++++++++++++++
 
+* added support for themes (fix issue #74)
 * show inline form errors in field title (fix issue #81)
 * fixed bugs in demo application
 * update to newest Bootstrap (fix issue #83)